+++ src/full/Agda/Syntax/Concrete.hs
@@ -389,9 +389,9 @@ data Declaration
   | Generalize Range [TypeSignature] -- ^ Variables to be generalized, can be hidden and/or irrelevant.
   | Field Range [FieldSignature]
   | FunClause LHS RHS WhereClause Bool
-  | DataSig     Range Induction Name [LamBinding] Expr -- ^ lone data signature in mutual block
-  | Data        Range Induction Name [LamBinding] Expr [TypeSignatureOrInstanceBlock]
-  | DataDef     Range Induction Name [LamBinding] [TypeSignatureOrInstanceBlock]
+  | DataSig     Range Name [LamBinding] Expr -- ^ lone data signature in mutual block
+  | Data        Range Name [LamBinding] Expr [TypeSignatureOrInstanceBlock]
+  | DataDef     Range Name [LamBinding] [TypeSignatureOrInstanceBlock]
   | RecordSig   Range Name [LamBinding] Expr -- ^ lone record signature in mutual block
   | RecordDef   Range Name (Maybe (Ranged Induction)) (Maybe HasEta) (Maybe (Name, IsInstance)) [LamBinding] [Declaration]
   | Record      Range Name (Maybe (Ranged Induction)) (Maybe HasEta) (Maybe (Name, IsInstance)) [LamBinding] Expr [Declaration]
@@ -745,9 +745,9 @@ instance HasRange Declaration where
   getRange (FieldSig _ _ x t)      = fuseRange x t
   getRange (Field r _)             = r
   getRange (FunClause lhs rhs wh _) = fuseRange lhs rhs `fuseRange` wh
-  getRange (DataSig r _ _ _ _)     = r
-  getRange (Data r _ _ _ _ _)      = r
-  getRange (DataDef r _ _ _ _)     = r
+  getRange (DataSig r _ _ _)       = r
+  getRange (Data r _ _ _ _)        = r
+  getRange (DataDef r _ _ _)       = r
   getRange (RecordSig r _ _ _)     = r
   getRange (RecordDef r _ _ _ _ _ _) = r
   getRange (Record r _ _ _ _ _ _ _)  = r
@@ -883,9 +883,9 @@ instance KillRange Declaration where
   killRange (Generalize r ds )      = killRange1 (Generalize noRange) ds
   killRange (Field r fs)            = killRange1 (Field noRange) fs
   killRange (FunClause l r w ca)    = killRange4 FunClause l r w ca
-  killRange (DataSig _ i n l e)     = killRange4 (DataSig noRange) i n l e
-  killRange (Data _ i n l e c)      = killRange4 (Data noRange i) n l e c
-  killRange (DataDef _ i n l c)     = killRange3 (DataDef noRange i) n l c
+  killRange (DataSig _ n l e)       = killRange3 (DataSig noRange) n l e
+  killRange (Data _ n l e c)        = killRange4 (Data noRange) n l e c
+  killRange (DataDef _ n l c)       = killRange3 (DataDef noRange) n l c
   killRange (RecordSig _ n l e)     = killRange3 (RecordSig noRange) n l e
   killRange (RecordDef _ n mi mb mn k d) = killRange6 (RecordDef noRange) n mi mb mn k d
   killRange (Record _ n mi mb mn k e d)  = killRange7 (Record noRange) n mi mb mn k e d
@@ -1095,9 +1095,9 @@ instance NFData Declaration where
   rnf (Generalize _ a)        = rnf a
   rnf (Field _ fs)            = rnf fs
   rnf (FunClause a b c d)     = rnf a `seq` rnf b `seq` rnf c `seq` rnf d
-  rnf (DataSig _ a b c d)     = rnf a `seq` rnf b `seq` rnf c `seq` rnf d
-  rnf (Data _ a b c d e)      = rnf a `seq` rnf b `seq` rnf c `seq` rnf d `seq` rnf e
-  rnf (DataDef _ a b c d)     = rnf a `seq` rnf b `seq` rnf c `seq` rnf d
+  rnf (DataSig _ a b c)       = rnf a `seq` rnf b `seq` rnf c
+  rnf (Data _ a b c d)        = rnf a `seq` rnf b `seq` rnf c `seq` rnf d
+  rnf (DataDef _ a b c)       = rnf a `seq` rnf b `seq` rnf c
   rnf (RecordSig _ a b c)     = rnf a `seq` rnf b `seq` rnf c
   rnf (RecordDef _ a b c d e f) = rnf a `seq` rnf b `seq` rnf c `seq` rnf d `seq` rnf e `seq` rnf f
   rnf (Record _ a b c d e f g)  = rnf a `seq` rnf b `seq` rnf c `seq` rnf d `seq` rnf e `seq` rnf f `seq` rnf g